@eleith: make require.js optimizable
@eleith:
when you passin the require module through the define function, i'm
unable to get r.js (require's node module for optimizing scripts) to
properly optimize beautify-html. my guess is that it does not read that
beautify-html has two dependencies.

however, when using define's first parameter option which is to pass an
array of dependency, r.js properly picks up the dependencies and bundles
them in the optimization process.