css-beautify: python cssbeautifier cli
This adds a command line tool for the python cssbeautifier.  It is based
on and re-uses as much code as possible from the js-beautify tool.

Long command line options make it possible to control all currently
implemented options in the CSS Beautifier.  Short command line options
are only supported for options that are identical to those of the
js-beautify tool.

Note that there is an inconsistency between the two tools caused by the
different default for the preserve_newlines option - True for
jsbeautifier and False for cssbeautifier, hence different command line
options --disable-preserve-newlines vs. --preserve-newlines.