Support wrapping script and style content
Content inside style and script tags can be wrapped in comment or
cdata to allow the text to contain characters and string that
would otherwise not be allowed in html.

Before this change the html beautifier would not see the wrappers.
This meant that it could incorrectly terminate script or style
tags. It also meant that the wrapper would be passed to child
beautifiers and they would have to handle them.

This change makes the html beautfier recognize the wrappers and
correctly handle formatting of them itself. The child beautifiers
see only the content not the wrapper.

Fixes #1641