issue #1683: fix switch case indentation
problem: Wrong indentation observed when the last line in a case is a right
brace.

solution: Deindent for case/default token is skipped whenever the previous
token is a right brace.
However this logic doesn't always work as highlighted in the issue.
Added code to skip deindent only if the closing right brace corresponds to
opening left brace of case action block.